+++ src/Stack/Types/BuildOpts.hs
@@ -140,6 +140,7 @@ defaultBuildOptsCLI = BuildOptsCLI
   , boptsCLIDryrun = False
   , boptsCLIFlags = Map.empty
   , boptsCLIGhcOptions = []
+  , boptsCLIProgsOptions = []
   , boptsCLIBuildSubset = BSAll
   , boptsCLIFileWatch = NoFileWatch
   , boptsCLIWatchAll = False
@@ -173,6 +174,7 @@ data BuildOptsCLI = BuildOptsCLI
   { boptsCLITargets :: ![Text]
   , boptsCLIDryrun :: !Bool
   , boptsCLIGhcOptions :: ![Text]
+  , boptsCLIProgsOptions :: ![(Text, [Text])]
   , boptsCLIFlags :: !(Map ApplyCLIFlag (Map FlagName Bool))
   , boptsCLIBuildSubset :: !BuildSubset
   , boptsCLIFileWatch :: !FileWatchOpts